+++ src/cardDb/resources/advancedResources.ts
@@ -1,16 +1,7 @@
-import { CardType, ResourceCard } from '@/types/cards';
+import { makeAdvancedResourceCard } from '@/factories/cards';
 import { EffectType, TargetTypes } from '@/types/effects';
 import { Resource } from '@/types/resources';
-import { Tokens } from '../units';
-
-export const makeAdvancedResourceCard = (
-    params: Omit<ResourceCard, 'cardType' | 'isUsed' | 'isAdvanced'>
-): ResourceCard => ({
-    cardType: CardType.RESOURCE,
-    isAdvanced: true,
-    isUsed: false,
-    ...params,
-});
+import { UTILITY_LANDS } from './utilityLands';
 
 // Duals - self ping duals
 const STRATOVOLCANO = makeAdvancedResourceCard({
@@ -304,93 +295,6 @@ const PREFECTURE_CAPITAL = makeAdvancedResourceCard({
     comesInTapped: true,
 });
 
-// Utility lands
-const COASTAL_CASTLE = makeAdvancedResourceCard({
-    name: 'Coastal Castle',
-    resourceType: Resource.WATER,
-    imgSrc: 'https://images.pexels.com/photos/4245826/pexels-photo-4245826.jpeg',
-    enterEffects: [
-        {
-            type: EffectType.HEAL,
-            strength: 2,
-            target: TargetTypes.ANY,
-        },
-    ],
-    comesInTapped: true,
-});
-
-const BAMBOO_FOOTBRIDGE = makeAdvancedResourceCard({
-    name: 'Bamboo Footbridge',
-    resourceType: Resource.BAMBOO,
-    imgSrc: 'https://images.pexels.com/photos/4090092/pexels-photo-4090092.jpeg',
-    enterEffects: [
-        {
-            type: EffectType.SUMMON_UNITS,
-            strength: 2,
-            summonType: Tokens.FROG,
-        },
-    ],
-    comesInTapped: true,
-});
-
-const SLAG_FIELDS = makeAdvancedResourceCard({
-    name: 'Slag Fields',
-    resourceType: Resource.IRON,
-    imgSrc: 'https://images.pexels.com/photos/955662/pexels-photo-955662.jpeg',
-    enterEffects: [
-        {
-            type: EffectType.BUFF_HAND_ATTACK,
-            strength: 1,
-        },
-    ],
-    comesInTapped: true,
-});
-
-const STARGAZERS_POINT = makeAdvancedResourceCard({
-    name: "Stargazer's Point",
-    resourceType: Resource.CRYSTAL,
-    imgSrc: 'https://images.pexels.com/photos/544268/pexels-photo-544268.jpeg',
-    enterEffects: [
-        {
-            type: EffectType.DISCARD_HAND,
-            strength: 1,
-            target: TargetTypes.SELF_PLAYER,
-        },
-        {
-            type: EffectType.DRAW,
-            strength: 1,
-        },
-    ],
-    comesInTapped: true,
-});
-
-const HOLY_TEMPLE = makeAdvancedResourceCard({
-    name: 'Holy Temple',
-    resourceType: Resource.FIRE,
-    imgSrc: 'https://images.pexels.com/photos/6710712/pexels-photo-6710712.jpeg',
-    enterEffects: [
-        {
-            type: EffectType.DEAL_DAMAGE,
-            strength: 1,
-            target: TargetTypes.ALL_OPPONENTS,
-        },
-    ],
-    comesInTapped: true,
-});
-
-const TREACHEROUS_DESERT = makeAdvancedResourceCard({
-    name: 'Treacherous Desert',
-    resourceType: Resource.GENERIC,
-    imgSrc: 'https://images.pexels.com/photos/273935/pexels-photo-273935.jpeg',
-    enterEffects: [
-        {
-            type: EffectType.DEAL_DAMAGE,
-            strength: 1,
-            target: TargetTypes.ALL_PLAYERS,
-        },
-    ],
-});
-
 export const AdvancedResourceCards = {
     // Duals - pain
     STRATOVOLCANO,
@@ -403,7 +307,6 @@ export const AdvancedResourceCards = {
     LUSH_REEF,
     HARBOR_TOWN,
     TANGLED_RUINS,
-    COASTAL_CASTLE,
     // Duals - heal
     LONE_TOWER,
     MYSTIFYING_LAKE,
@@ -416,9 +319,6 @@ export const AdvancedResourceCards = {
     CITY_CANALS,
     PREFECTURE_CAPITAL,
     // Utility - mono
-    BAMBOO_FOOTBRIDGE,
-    STARGAZERS_POINT,
-    SLAG_FIELDS,
-    HOLY_TEMPLE,
-    TREACHEROUS_DESERT,
+    ...UTILITY_LANDS,
 };
+export { makeAdvancedResourceCard } from '@/factories/cards';
